refactor(sync): Change Koin WorkManager dependency to api
This commit modifies the `core/sync/build.gradle.kts` file to expose the Koin WorkManager dependency to other modules.

The `io.insert.koin.androidx.workmanager` dependency is changed from an `implementation` configuration to an `api` configuration. This allows modules that depend on `core:sync` to also have access to the Koin WorkManager APIs.